# Use SIMD to speed up Paeth unfiltering for bpp=4, bpp=8