--- a/src/filter.rs
+++ b/src/filter.rs
@@ -9,9 +9,9 @@ use crate::common::BytesPerPixel;
 /// feature of Rust gets stabilized.
 #[cfg(feature = "unstable")]
 mod simd {
-    use std::simd::cmp::{SimdOrd, SimdPartialEq};
+    use std::simd::cmp::{SimdOrd, SimdPartialEq, SimdPartialOrd};
     use std::simd::num::{SimdInt, SimdUint};
-    use std::simd::{u8x4, u8x8, LaneCount, Simd, SupportedLaneCount};
+    use std::simd::{u8x4, u8x8, LaneCount, Simd, SimdElement, SupportedLaneCount};
 
     /// This is an equivalent of the `PaethPredictor` function from
     /// [the spec](http://www.libpng.org/pub/png/spec/1.2/PNG-Filters.html#Filter-type-4-Paeth)
@@ -53,26 +53,80 @@ mod simd {
             .select(a, smallest.simd_eq(pb).select(b, c))
     }
 
-    /// Memory of previous pixels (as needed to unfilter `FilterType::Paeth`).
-    /// See also https://www.w3.org/TR/png/#filter-byte-positions
-    #[derive(Default)]
-    struct PaethState<const N: usize>
+    /// Equivalent to `simd::paeth_predictor` but does not temporarily convert
+    /// the SIMD elements to `i16`.
+    fn paeth_predictor_u8<const N: usize>(
+        a: Simd<u8, N>,
+        b: Simd<u8, N>,
+        c: Simd<u8, N>,
+    ) -> Simd<u8, N>
     where
         LaneCount<N>: SupportedLaneCount,
     {
+        // Calculates the absolute difference between `a` and `b`.
+        fn abs_diff_simd<const N: usize>(a: Simd<u8, N>, b: Simd<u8, N>) -> Simd<u8, N>
+        where
+            LaneCount<N>: SupportedLaneCount,
+        {
+            a.simd_max(b) - b.simd_min(a)
+        }
+
+        // Uses logic from `filter::filter_paeth` to calculate absolute values
+        // entirely in `Simd<u8, N>`. This method avoids unpacking and packing
+        // penalties resulting from conversion to and from `Simd<i16, N>`.
+        // ```
+        //     let pa = b.max(c) - c.min(b);
+        //     let pb = a.max(c) - c.min(a);
+        //     let pc = if (a < c) == (c < b) {
+        //         pa.max(pb) - pa.min(pb)
+        //     } else {
+        //         255
+        //     };
+        // ```
+        let pa = abs_diff_simd(b, c);
+        let pb = abs_diff_simd(a, c);
+        let pc = a
+            .simd_lt(c)
+            .simd_eq(c.simd_lt(b))
+            .select(abs_diff_simd(pa, pb), Simd::splat(255));
+
+        let smallest = pc.simd_min(pa.simd_min(pb));
+
+        // Paeth algorithm breaks ties favoring a over b over c, so we execute the following
+        // lane-wise selection:
+        //
+        //     if smalest == pa
+        //         then select a
+        //         else select (if smallest == pb then select b else select c)
+        smallest
+            .simd_eq(pa)
+            .select(a, smallest.simd_eq(pb).select(b, c))
+    }
+
+    /// Memory of previous pixels (as needed to unfilter `FilterType::Paeth`).
+    /// See also https://www.w3.org/TR/png/#filter-byte-positions
+    #[derive(Default)]
+    struct PaethState<T, const N: usize>
+    where
+        T: SimdElement,
+        LaneCount<N>: SupportedLaneCount,
+    {
         /// Previous pixel in the previous row.
-        c: Simd<i16, N>,
+        c: Simd<T, N>,
 
         /// Previous pixel in the current row.
-        a: Simd<i16, N>,
+        a: Simd<T, N>,
     }
 
     /// Mutates `x` as needed to unfilter `FilterType::Paeth`.
     ///
     /// `b` is the current pixel in the previous row.  `x` is the current pixel in the current row.
     /// See also https://www.w3.org/TR/png/#filter-byte-positions
-    fn paeth_step<const N: usize>(state: &mut PaethState<N>, b: Simd<u8, N>, x: &mut Simd<u8, N>)
-    where
+    fn paeth_step<const N: usize>(
+        state: &mut PaethState<i16, N>,
+        b: Simd<u8, N>,
+        x: &mut Simd<u8, N>,
+    ) where
         LaneCount<N>: SupportedLaneCount,
     {
         // Storing the inputs.
@@ -85,6 +139,24 @@ mod simd {
         // Preparing for the next step.
         state.c = b;
         state.a = x.cast::<i16>();
+    }
+
+    /// Computes the Paeth predictor without converting `u8` to `i16`.
+    ///
+    /// See `simd::paeth_step`.
+    fn paeth_step_u8<const N: usize>(
+        state: &mut PaethState<u8, N>,
+        b: Simd<u8, N>,
+        x: &mut Simd<u8, N>,
+    ) where
+        LaneCount<N>: SupportedLaneCount,
+    {
+        // Calculating the new value of the current pixel.
+        *x += paeth_predictor_u8(state.a, b, state.c);
+
+        // Preparing for the next step.
+        state.c = b;
+        state.a = *x;
     }
 
     fn load3(src: &[u8]) -> u8x4 {
@@ -100,7 +172,7 @@ mod simd {
         debug_assert_eq!(prev_row.len(), curr_row.len());
         debug_assert_eq!(prev_row.len() % 3, 0);
 
-        let mut state = PaethState::<4>::default();
+        let mut state = PaethState::<i16, 4>::default();
         while prev_row.len() >= 4 {
             // `u8x4` requires working with `[u8;4]`, but we can just load and ignore the first
             // byte from the next triple.  This optimization technique mimics the algorithm found
@@ -126,6 +198,30 @@ mod simd {
         store3(x, curr_row);
     }
 
+    /// Undoes `FilterType::Paeth` for `BytesPerPixel::Four` and `BytesPerPixel::Eight`.
+    ///
+    /// This function calculates the Paeth predictor entirely in `Simd<u8, N>`
+    /// without converting to an intermediate `Simd<i16, N>`. Doing so avoids
+    /// paying a small performance penalty converting between types.
+    pub fn unfilter_paeth_u8<const N: usize>(prev_row: &[u8], curr_row: &mut [u8])
+    where
+        LaneCount<N>: SupportedLaneCount,
+    {
+        debug_assert_eq!(prev_row.len(), curr_row.len());
+        debug_assert_eq!(prev_row.len() % N, 0);
+        assert!(matches!(N, 4 | 8));
+
+        let mut state = PaethState::<u8, N>::default();
+        for (prev_row, curr_row) in prev_row.chunks_exact(N).zip(curr_row.chunks_exact_mut(N)) {
+            let b = Simd::from_slice(prev_row);
+            let mut x = Simd::from_slice(curr_row);
+
+            paeth_step_u8(&mut state, b, &mut x);
+
+            curr_row[..N].copy_from_slice(&x.to_array()[..N]);
+        }
+    }
+
     fn load6(src: &[u8]) -> u8x8 {
         u8x8::from_array([src[0], src[1], src[2], src[3], src[4], src[5], 0, 0])
     }
@@ -139,7 +235,7 @@ mod simd {
         debug_assert_eq!(prev_row.len(), curr_row.len());
         debug_assert_eq!(prev_row.len() % 6, 0);
 
-        let mut state = PaethState::<8>::default();
+        let mut state = PaethState::<i16, 8>::default();
         while prev_row.len() >= 8 {
             // `u8x8` requires working with `[u8;8]`, but we can just load and ignore the first two
             // bytes from the next pixel.  This optimization technique mimics the algorithm found
@@ -678,23 +774,34 @@ pub(crate) fn unfilter(
                     }
                 }
                 BytesPerPixel::Four => {
-                    let mut a_bpp = [0; 4];
-                    let mut c_bpp = [0; 4];
-                    for (chunk, b_bpp) in current.chunks_exact_mut(4).zip(previous.chunks_exact(4))
+                    #[cfg(feature = "unstable")]
+                    simd::unfilter_paeth_u8::<4>(previous, current);
+
+                    #[cfg(not(feature = "unstable"))]
                     {
-                        let new_chunk = [
-                            chunk[0]
-                                .wrapping_add(filter_paeth_decode(a_bpp[0], b_bpp[0], c_bpp[0])),
-                            chunk[1]
-                                .wrapping_add(filter_paeth_decode(a_bpp[1], b_bpp[1], c_bpp[1])),
-                            chunk[2]
-                                .wrapping_add(filter_paeth_decode(a_bpp[2], b_bpp[2], c_bpp[2])),
-                            chunk[3]
-                                .wrapping_add(filter_paeth_decode(a_bpp[3], b_bpp[3], c_bpp[3])),
-                        ];
-                        *TryInto::<&mut [u8; 4]>::try_into(chunk).unwrap() = new_chunk;
-                        a_bpp = new_chunk;
-                        c_bpp = b_bpp.try_into().unwrap();
+                        let mut a_bpp = [0; 4];
+                        let mut c_bpp = [0; 4];
+                        for (chunk, b_bpp) in
+                            current.chunks_exact_mut(4).zip(previous.chunks_exact(4))
+                        {
+                            let new_chunk = [
+                                chunk[0].wrapping_add(filter_paeth_decode(
+                                    a_bpp[0], b_bpp[0], c_bpp[0],
+                                )),
+                                chunk[1].wrapping_add(filter_paeth_decode(
+                                    a_bpp[1], b_bpp[1], c_bpp[1],
+                                )),
+                                chunk[2].wrapping_add(filter_paeth_decode(
+                                    a_bpp[2], b_bpp[2], c_bpp[2],
+                                )),
+                                chunk[3].wrapping_add(filter_paeth_decode(
+                                    a_bpp[3], b_bpp[3], c_bpp[3],
+                                )),
+                            ];
+                            *TryInto::<&mut [u8; 4]>::try_into(chunk).unwrap() = new_chunk;
+                            a_bpp = new_chunk;
+                            c_bpp = b_bpp.try_into().unwrap();
+                        }
                     }
                 }
                 BytesPerPixel::Six => {
@@ -735,31 +842,46 @@ pub(crate) fn unfilter(
                     }
                 }
                 BytesPerPixel::Eight => {
-                    let mut a_bpp = [0; 8];
-                    let mut c_bpp = [0; 8];
-                    for (chunk, b_bpp) in current.chunks_exact_mut(8).zip(previous.chunks_exact(8))
+                    #[cfg(feature = "unstable")]
+                    simd::unfilter_paeth_u8::<8>(previous, current);
+
+                    #[cfg(not(feature = "unstable"))]
                     {
-                        let new_chunk = [
-                            chunk[0]
-                                .wrapping_add(filter_paeth_decode(a_bpp[0], b_bpp[0], c_bpp[0])),
-                            chunk[1]
-                                .wrapping_add(filter_paeth_decode(a_bpp[1], b_bpp[1], c_bpp[1])),
-                            chunk[2]
-                                .wrapping_add(filter_paeth_decode(a_bpp[2], b_bpp[2], c_bpp[2])),
-                            chunk[3]
-                                .wrapping_add(filter_paeth_decode(a_bpp[3], b_bpp[3], c_bpp[3])),
-                            chunk[4]
-                                .wrapping_add(filter_paeth_decode(a_bpp[4], b_bpp[4], c_bpp[4])),
-                            chunk[5]
-                                .wrapping_add(filter_paeth_decode(a_bpp[5], b_bpp[5], c_bpp[5])),
-                            chunk[6]
-                                .wrapping_add(filter_paeth_decode(a_bpp[6], b_bpp[6], c_bpp[6])),
-                            chunk[7]
-                                .wrapping_add(filter_paeth_decode(a_bpp[7], b_bpp[7], c_bpp[7])),
-                        ];
-                        *TryInto::<&mut [u8; 8]>::try_into(chunk).unwrap() = new_chunk;
-                        a_bpp = new_chunk;
-                        c_bpp = b_bpp.try_into().unwrap();
+                        let mut a_bpp = [0; 8];
+                        let mut c_bpp = [0; 8];
+                        for (chunk, b_bpp) in
+                            current.chunks_exact_mut(8).zip(previous.chunks_exact(8))
+                        {
+                            let new_chunk = [
+                                chunk[0].wrapping_add(filter_paeth_decode(
+                                    a_bpp[0], b_bpp[0], c_bpp[0],
+                                )),
+                                chunk[1].wrapping_add(filter_paeth_decode(
+                                    a_bpp[1], b_bpp[1], c_bpp[1],
+                                )),
+                                chunk[2].wrapping_add(filter_paeth_decode(
+                                    a_bpp[2], b_bpp[2], c_bpp[2],
+                                )),
+                                chunk[3].wrapping_add(filter_paeth_decode(
+                                    a_bpp[3], b_bpp[3], c_bpp[3],
+                                )),
+                                chunk[4].wrapping_add(filter_paeth_decode(
+                                    a_bpp[4], b_bpp[4], c_bpp[4],
+                                )),
+                                chunk[5].wrapping_add(filter_paeth_decode(
+                                    a_bpp[5], b_bpp[5], c_bpp[5],
+                                )),
+                                chunk[6].wrapping_add(filter_paeth_decode(
+                                    a_bpp[6], b_bpp[6], c_bpp[6],
+                                )),
+                                chunk[7].wrapping_add(filter_paeth_decode(
+                                    a_bpp[7], b_bpp[7], c_bpp[7],
+                                )),
+                            ];
+                            *TryInto::<&mut [u8; 8]>::try_into(chunk).unwrap() = new_chunk;
+                            a_bpp = new_chunk;
+                            c_bpp = b_bpp.try_into().unwrap();
+                        }
                     }
                 }
             }