- Add SIMD-accelerated unfiltering for `BytesPerPixel::{Four, Eight}`
- Add `paeth_predictor_u8`, `paeth_step_u8`, and `unfilter_paeth_u8` for calculating the Paeth predictor entirely in `Simd<u8, N>` without converting to `Simd<i16, N>`
- Parameterize `PaethState` by `T`, the `SimdElement` type

Calculating the filter entirely in `u8` avoids penalties which result from unpacking to and packing from `Simd<i16, N>`.
`pa`, `pb`, and `pc` can all be calculated using absolute differences and reusing the logic of `filter::filter_paeth`.

---

This PR duplicates the existing logic for unfilter3 and unfilter6. I tried to apply this approach to those cases but it was not a complete win: staying in `u8` was faster for the SSE2 baseline profile but slower for AVX2.

Benchmark numbers
```bash
# 1. bpp=4, autovectorization
$ RUSTFLAGS="-Ctarget-cpu=x86-64-v3" cargo +nightly bench --features=benchmarks -- filter=Paeth/bpp=4            
                        time:   [12.895 µs 12.900 µs 12.905 µs]
                        thrpt:  [1.1824 GiB/s 1.1829 GiB/s 1.1833 GiB/s]

# 2. bpp=4, portable_simd
$ RUSTFLAGS="-Ctarget-cpu=x86-64-v3" cargo +nightly bench --features=benchmarks,unstable -- filter=Paeth/bpp=4
                        time:   [11.143 µs 11.162 µs 11.195 µs]
                        thrpt:  [1.3630 GiB/s 1.3671 GiB/s 1.3694 GiB/s]

# 3. bpp=8, autovectorization
$ RUSTFLAGS="-Ctarget-cpu=x86-64-v3" cargo +nightly bench --features=benchmarks -- filter=Paeth/bpp=8
                        time:   [18.209 µs 18.226 µs 18.246 µs]
                        thrpt:  [1.6725 GiB/s 1.6744 GiB/s 1.6760 GiB/s]

# 4. bpp=8, portable_simd
$ RUSTFLAGS="-Ctarget-cpu=x86-64-v3" cargo +nightly bench --features=benchmarks,unstable -- filter=Paeth/bpp=8
                        time:   [11.172 µs 11.180 µs 11.187 µs]
                        thrpt:  [2.7279 GiB/s 2.7296 GiB/s 2.7316 GiB/s]
```