@HeroicKatora 2024-08-15
Nicely spotted.

---

Aside: I'm interpreting this right, it's (apparently) not memory throughput limited?  That makes me wonder if one could speed up some of the other predictors by scanning multiple pages in parallel. Of course not Paeth which has the whole dependency chain that's hard to reorder. Although.. Well it would require rewriting the whole scanline tracking which would be too hard to maintain just the sake for an unstable partial feature.